I benchmarked using this command:
`GRPC_BENCHMARK_WARMUP=20s GRPC_BENCHMARK_DURATION=40s ./bench.sh scala_fs2_bench`

ORIGINAL(fs2-grpc 1.1.5)
```
-----------------------------------------------------------------------------------------------------------------------------------------
| name                        |   req/s |   avg. latency |        90 % in |        95 % in |        99 % in | avg. cpu |   avg. memory |
-----------------------------------------------------------------------------------------------------------------------------------------
| scala_fs2                   |    2558 |      316.62 ms |      572.77 ms |      786.95 ms |         1.01 s |  106.58% |    221.16 MiB |
-----------------------------------------------------------------------------------------------------------------------------------------
```

fs2-grpc main
3d63db8d8a71e9a5102848f90e578841700c0718
```
-----------------------------------------------------------------------------------------------------------------------------------------
| name                        |   req/s |   avg. latency |        90 % in |        95 % in |        99 % in | avg. cpu |   avg. memory |
-----------------------------------------------------------------------------------------------------------------------------------------
| scala_fs2                   |    2202 |      405.11 ms |      641.56 ms |      864.58 ms |         1.33 s |  103.79% |    229.26 MiB |
-----------------------------------------------------------------------------------------------------------------------------------------
```

Channel instead of ce Queue
69b855f17309484eab51bcebe2ba9056c29adaa9
```
-----------------------------------------------------------------------------------------------------------------------------------------
| name                        |   req/s |   avg. latency |        90 % in |        95 % in |        99 % in | avg. cpu |   avg. memory |
-----------------------------------------------------------------------------------------------------------------------------------------
| scala_fs2                   |    2161 |      410.97 ms |      669.56 ms |      844.99 ms |         1.61 s |  104.27% |    232.76 MiB |
-----------------------------------------------------------------------------------------------------------------------------------------
```

For now every update makes situation worse