@ahjohannessen 2022-01-16
I think `StreamIngest` is not used in the benchmarks given that:

```
service Greeter {
  // Sends a greeting
  rpc SayHello (HelloRequest) returns (HelloReply) {}
}
```

is unary / unary. So that is a bit odd.