@nikiforo 2022-01-16
unary listener without fs2.Stream
2390ad79103baa842ac1f770c54a50bec35bb24a
```
-----------------------------------------------------------------------------------------------------------------------------------------
| name                        |   req/s |   avg. latency |        90 % in |        95 % in |        99 % in | avg. cpu |   avg. memory |
-----------------------------------------------------------------------------------------------------------------------------------------
| scala_fs2                   |    2239 |      390.23 ms |      681.94 ms |      762.23 ms |         1.52 s |  104.82% |     236.3 MiB |
-----------------------------------------------------------------------------------------------------------------------------------------
```