@ahjohannessen 2022-01-16
@nikiforo This could perhaps be of interest: https://github.com/typelevel/fs2-grpc/pull/394