@djspiewak 2022-01-21
Trying to capture some thoughts…

The `unsafeRunSync`s are almost certainly the largest source of trouble here. Reading through `ServerCall`'s javadoc a bit, it doesn't seem like we're *required* to hard-block the calling thread, merely allowed to do so. Removing the blocking altogether would pretty dramatically improve throughput and lower context shift penalties. In the one or two places where the code is relying on sequentialization due to blocking, you can reestablish that by wrapping the sequentially-next action in `Sync[F].delay(...)` and just `flatMap`ping with the dispatched action.

Once you've done this, you should be able to reduce the threads used by io-grpc down to an absolute minim. The ideal (in my measurements) is somewhere around 1-2 threads, because you're looking at something which is purely IO-bound and just responsible for grabbing data and shoveling it off into the compute pool. This will also reduce a lot of the context shift penalties since thread affinity will improve a lot.

Finally, you need to find a way to reestablish backpressure guarantees that have been lost by making the listener non-blocking. I'm not sure exactly how to do this in io-grpc, but the documentation suggests that there should be a way given that they allow the listeners to be non-blocking in the first place. What you're looking for is basically something like Netty's `setAutoRead` and `read()`, which controls the paging of the next block of bytes from the underlying IO system and allows you to explicitly track and control when it's safe to do exactly that.

On the multiple call handler, it seems most likely that a higher performance `Queue` would help alleviate a major bottleneck, since CE's `Queue` right now is about an order of magnitude slower than it strictly needs to be.