@ahjohannessen 2022-01-22
@djspiewak I did an experiment by redoing the unary to unary call [here](https://github.com/typelevel/fs2-grpc/pull/486/files#).