@ahjohannessen 2022-02-27
I am closing this as performance work by @naoh87 that improves this has been merged into main.